deprecate main.js concept and directly define main() in each entrypoint
