+++ web/player/song-select.js
@@ -1,0 +1,11 @@
+import { createApp } from "vue";
+
+function main() {}
+
+const app = createApp({
+    setup() {
+        return main();
+    },
+});
+
+app.mount("#app");
