+++ web/utils/commit-info.js
@@ -1,1 +1,1 @@
-export const COMMIT_ID = '578fa46575be'
+export const COMMIT_ID = '05424a7f1f19'
